Add rasterizer 2D texture filtering /w 6xXBRZ
This implements texture filtering (initally via xBRZ) to individual
textures, which are cached in the regular rasterizer cache. This means
that the final image will not have expensive filtering operations applied
to it, while onscreen assets (fonts, sprites, textures applied to 3D
models, etc) will be much nicer at higher resolutions.

This also implements intelligent filtering, in that the filtering
implementation will ignore frequently updated textures, and ones that
would otherwise grind this to a halt.

Other texture formats can be implemented with this, via adding a
FilteringType and a implementation in the filterTexture() method.